+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (**frontend**): Serve markdown for marketing & docs pages to agents by @anupcowkur in #3621
 - (**frontend**): Add onboarding flow by @anupcowkur in #3613
 - (**frontend**): Make dashboard header sticky on scroll by @anupcowkur in #3489
+- (**kmp**): Implement KMP sdk (#3503) by @abhaysood in #3503
 
 ### :bug: Bug fixes
 
@@ -72,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (**deps**): Bump github.com/jackc/pgx/v5 in /backend/symboloader (#3521) by @dependabot[bot] in #3521
 - (**deps**): Bump github.com/jackc/pgx/v5 in /backend/testinfra (#3520) by @dependabot[bot] in #3520
 - (**deps**): Bump github.com/jackc/pgx/v5 in /self-host/sessionator (#3519) by @dependabot[bot] in #3519
+- (**frontend**): Use c15t cookie banner by @anupcowkur in #3640
 - (**frontend**): Improve line height comment by @anupcowkur in #3635
 - (**frontend**): Improve line heights by @anupcowkur
 - (**frontend**): Update header star count by @anupcowkur in #3634